+++ docs/CHANGELOG.md
@@ -5,12 +5,13 @@
 ## 6.2.0
 - bumped version up to be in sync with mycms-frontend-commons, mycms-commons
 - interpret parameter loadDetailsMode of GenericSearchOptions on server-request
+- improved pdf-generation
 
 ### new features
 - none
 
 ### improvements
-- none
+- improved pdf-generation
 
 ### bug fixes
 - none